category utils tests reorganised and more created
